Change divceil to operate on size_t
All the uses of this function pass in a size_t, which is an unsigned
long on GCC. On 32 bit systems this is the same as an unsigned int, but
on 64 bit systems it's not. Just use size_t to ensure there are no
overflows or truncations.

+++ base32.c
@@ -31,8 +31,8 @@ new_z(const size_t len)
 }
 
 /* Functions from zutil */
-static unsigned int
-divceil(unsigned int n, unsigned int d)
+static size_t
+divceil(size_t n, size_t d)
 {
 	return n/d+((n%d)!=0);
 }
